+++ noxfile.py
@@ -8,7 +8,7 @@ if typing.TYPE_CHECKING:
 nox.options.default_venv_backend = "uv|virtualenv"
 
 
-@nox.session(python=["3.13"])
+@nox.session(python=["2.12", "3.13"])
 def tests(session: "Session"):
     session.run(
         "uv",
@@ -24,7 +24,7 @@ def tests(session: "Session"):
     )
 
 
-@nox.session(python=["3.13"], tags=["coverage"])
+@nox.session(python=["3.12", "3.13"], tags=["coverage"])
 def tests_with_coverage(session: "Session"):
     session.env["COVERAGE_FILE"] = f".coverage.{session.python}"
     session.run(